Fix many bugs, add many features:
* Fix functions not returning values correctly:

Since LISP uses implicit returns, we need to allow for functions with
expressions to return their values, but also run other code as needed.
To acheive this, we (ab)use the comma operator, which will evaluate
each of the expressions, and return the last expression in the comma
list as the result of the expression.

* Modify AST parser/code generator to allow for nested expressions

* Handle exceptions better in main loop

* Add exit helper function

* Dont do anything if line is empty

* Add doc comments

* Add list related functions

* Add set function